+++ tests/e2e/cases/generated_c_dce_class/in.ts
@@ -6,6 +6,18 @@ class UsedBox {
     }
 
     read(): number {
+        class DeadLocalBox {
+            value: number;
+
+            constructor(value: number) {
+                this.value = value;
+            }
+
+            read(): number {
+                return this.value;
+            }
+        }
+
         return this.value + 2;
     }
 }